The Big Auth Rewrite: now with multiple account support
This change adds multiple account support to pillbug.
To accomplish that, I rewrote the authentication context to be significantly more convenient to deal with.

Along the way I added some generic helpers:
- `store-backed.ts`: A base class backed by an ephemeral Store. (and optionally persistent Store, too) The new SessionAuthManager is built on top of this.
- `blocking-load.tsx` and `tracked-blocking-load.tsx`: I needed a way to stop the frame from trying to load until a megalodon client had been created. This is a bit overbuilt for that purpose, but will be handy anywhere we need to block ui and wait for one or more async resources to finish (fetching share chains seems like a good candidate)

I also added an initial settings model and page, and implemented a toggle for internet time (aka beats).

Unfortunately due to changes in how logged in accounts are stored, you'll have to log in again. Sorry! I don't expect that to happen again.